requirements.txt
streamlit
pandas
scikit-learn
joblib
requests
python-dotenv
plotly